Bug 65238: detect and report number of PFs on the port
There is number of PFs on the NIC in the EFX configuration, but
there is no number of PFs on the port. This patch adds it.
This counter is useful for finding out if there are unprivileged
functions that are assigned to the port. For unprivileged functions
themselves, this count is 0.

Signed-off-by: Roman Zhukov <Roman.Zhukov@oktetlabs.ru>