net/sfc/base: add support for the NO_CONT_EV RX event mode
The recently added NO_CONT_EV mode is recommended for when looking for maximum
throughput on 100G links as it allows the firmware to operate more
efficiently. The biggest benefit is when using scatter and jumbo frames, but
it is also necessary to achieve line rate in other cases.

Support for NO_CONT_EV when scatter is disabled is simple - the main datapth
change is to always read the packet length from the prefix, not the
event. This requires storing a flag with the event queue so the event handler
knows NO_CONT_EV mode is in use.

Supporting NO_CONT_EV with scatter would require an API change.

(Now the ee_flags field in efx_evq_t is used on the datapath, move it next to
other fields which are also read on the datapath to try to avoid reading an
additional cache line.)

Signed-off-by: Mark Spender <mspender@solarflare.com>