common/sfc_efx/base: read EF100 design parameters
Read the design parameter TLV in the Riverhead NIC initialization
routine after VI windows allocation and store the values of required
design parameters in the NIC configuration structure. If the value is
not specified in design parameter TLV, the default value is used.

For more details, see SF-119689-TC "EF100 Host Interface".

Signed-off-by: Igor Romanov <igor.romanov@oktetlabs.ru>