common/sfc_efx/base: NOCOMMIT: add switchdev mode API
It is a temporary workaround for the EVB infrastracture being
completely unavailable when MAE rules are inserted, particularly,
efx_mcdi_vadapter_set_mac(), which is omitted with that change.

Add an API to toggle the default EVB mode or switchdev mode.
In EVB mode, FW provides interface to control bridging inside NIC in
terms of vSwtich/vPort/vAdaptor. In switchdev mode MAE action rules
should be used to route traffic inside NIC.

Signed-off-by: Igor Romanov <igor.romanov@oktetlabs.ru>