net/sfc: support count action for the flow API
The implementation allows setting a single counter per flow
rule, shared or dedicated.

To get counter values, the driver needs to start counter stream on
an RxQ, to which counter packets are delivered. Driver creates a
dedicated for counters RxQ, starts the counter stream on it and polls
it periodically utilizing the first available service core. Therefore
user has to specify at least one service core for counters to work.
A single service core is shared by all devices that use the sfc driver
and create a flow rule with count action.

Signed-off-by: Igor Romanov <igor.romanov@oktetlabs.ru>