runtime: add switch to prevent freeing of already allocated hugepages
The scripts might not solve all hugepages related problems and users
might have other allocation mechanisms in place. In fact that is even
suggested in the comments of the config file, should setups become more
complex.

In such complex setups it might not always be correct to reduce the
number of hugepages. So create a way to control that and keep the old
default.

Signed-off-by: Henning Schild <henning.schild@siemens.com>