net/virtio: wait for reset completion
Waiting is required when HW supports virtio and it requries
time to perform reset.

Virtio spec [1] is vague in this place since init sequence in
section 3.1.1 does not mention that a driver must wait for reset
completion explicitly.

The behaviour to return non-zero status value is in line with
section 4.1.4.3.1 of the document:
    The device MUST reset when 0 is written to device_status, and
    present a 0 in device_status once that is done.

Linux virtio driver is wating for reset completion in a similar way.

[1] https://docs.oasis-open.org/virtio/virtio/v1.1/csprd01/virtio-v1.1-csprd01.html

Suggested-by: Aleksandar Radovanovic <radovano@xilinx.com>
Signed-off-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>