common/sfc_efx/base: support match on recirculation ID
Currently, libefx only supports match on outer rule ID,
when an action rule is linked with an exact outer rule.

At the same time, in MAE, an outer rule (when hit) can
provide a recirculation ID for a packet to action rule
lookup. An action rule can match on this ID instead of
matching the an outer rule ID. This allows one to link
a single action rule with multiple outer rules sharing
the same recirculation ID. Also, one may link multiple
action rules matching on a given recirculation ID with
a single outer rule bearing this ID in order to have a
means to remove and insert back the outer rule without
the need to re-insert the action rules, which would be
needed if they matched on the outer rule (resource) ID.

Introduce required means to support such functionality.

Signed-off-by: Ivan Malov <ivan.malov@oktetlabs.ru>