net/sfc: prepare Rx packet mark mechanism for tunnel offload
As per the offload model, for a given tunnel endpoint, there
is a primary flow associated with a table of secondary flows.
The primary flow matches on the encapsulation header, strips
it and jumps to the associated table. A flow in the table is
supposed to match on a narrow set of network fields and pass
matching packets to some final destination, in example, to a
VF. Packets which miss in the table must hit the admin's PMD.
The PMD may be eventually asked to recover tunnel data for a
missed packet. The encapsulation header may also be restored.

In this particular implementation, the primary flow does not
strip the encapsulation header and it is not really bound to
the table flows, which perform the entire offload themselves.
Its only purpose is to catch packets which miss in the table;
such packets need to be marked to let the PMD recognise them.

To prepare for later patches, add basic model definitions as
well as describe the mark sub-division. Fence off a sensible
amount of high bits in the mark to meet tunnel offload needs.

Signed-off-by: Ivan Malov <ivan.malov@oktetlabs.ru>