net/sfc: support MAE-based secondary flows in tunnel offload
As per the offload structure, a secondary flow (one may have
multiple of such) provides narrower match for packets coming
from the primary flow and takes them to a specific recipient.

In this implementation, the primary flow is not chained with
the secondary ones; it is solely used to mark missed packets.
The secondary flow takes encapsulated packets right from the
wire and has to decapsulate the matching ones prior delivery.

The secondary flow is MAE-based. It is represented by a pair
of rules, an outer rule (OR) and an action rule (AR). The OR
is the same across the whole group of secondary flows in the
given tunnel entry meaning the outer match criteria coincide.

Define more parts of tunnel offload support and provide code
to recognise secondary (MAE-based) flows and put them to use.

Signed-off-by: Ivan Malov <ivan.malov@oktetlabs.ru>