net/sfc: support VNRX-backed primary flows in tunnel offload
As per the offload structure, a primary flow installed by an
application contains PMD-specific actions and an action JUMP,
which chains the flow with a table (group) of secondary ones.

In this particular implementation, there is no real chaining
from the HW standpoint, and the primary flow is only used to
catch packets which miss on secondary flow table lookup. The
table flows are backed by Match-Action Engine (MAE). Packets
which miss in MAE go to admin's PF and hit the VNIC Rx block.
The primary flow is a filter sitting in this block. It marks
the missed packets and then directs them to default Rx queue.

Define more parts of tunnel offload support and provide code
for detecting, parsing and tracking the primary (VNRX) flows.

Signed-off-by: Ivan Malov <ivan.malov@oktetlabs.ru>