net/sfc: provide support for flow counters in tunnel offload
Applications will be able to query counters in primary flows
and also in secondary flows. The former will only report the
number of hits, which is actually a sum of two contributions.
The first one is based on counting missed packets seen by Rx
burst method (identified by the tunnel mark). The second one
is based on indirect increments made by MAE counter-on-queue
packet processing service for the associated secondary flows.

Signed-off-by: Ivan Malov <ivan.malov@oktetlabs.ru>