[2/6]net/virtio: fix Rx checksum offload for TCP case
Fix checksum algorithm to perform final complement even
if the sum is 0xffff, since specification doesn't say it
must not be performed and, for example, in case of TCP a
legit zero value is obtained. Revert the obtained checksum
to 0xffff if it's zero in case of UDP according to RFC 768.

Fixes: 96cb6711939 ("net/virtio: support Rx checksum offload")

Signed-off-by: Ivan Ilchenko <ivan.ilchenko@oktetlabs.ru>