[6/6]net/virtio: remove comment on SCTP in Rx checksum offload
Remove a comment the checksum algorithm is wrong in case of
SCTP because SCTP uses not ones' complement but CRC. The comment
is unnecessary since device guarantees that ones' complement
must be used to obtain a correctly checksummed packet, e.g. in
case of SCTP the VIRTIO_NET_HDR_F_NEEDS_CSUM is not set and the
algorithm is not executed anyway.

Fixes: 96cb6711939 ("net/virtio: support Rx checksum offload")

Signed-off-by: Ivan Ilchenko <ivan.ilchenko@oktetlabs.ru