examples/vhost: fix port init failure in mergeable mode
When the example starts in mergeable mode with an i40e port,
it fails to launch because the examples use default mtu MAX_MTU
to configure ethdev. The root cause is some devices have Ethernet
frame overhead and then MAX_MTU will be larger than device's max
mtu, so the ethdev configure will fail.

This patch checks the device's max mtu before setting the ethdev
configuration. If the device has a max mtu, use that value to
configure.

Fixes: 1bb4a528c41f ("ethdev: fix max Rx packet length")

Signed-off-by: Chenbo Xia <chenbo.xia@intel.com>
Reported-by: Xingguang He <xingguang.he@intel.com>
Reviewed-by: Maxime Coquelin <maxime.coquelin@redhat.com>